Data Extractions and the Datasets
In k_fold_valid_extraction.py, the implementation of k-fold
extraction is completed, and the generated datasets are in the
folder of process/k_fold/, there are 2 types of datasets, one is
the combined without randomization, and the other is the combined
with randomization. The randomized ones have Random at the last
in the file name.